Updated to instant command
