--- src/main/java/frc/robot/commands/arm/ArmAuto.java
@@ -1,47 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands.arm;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.arm.ArmSubsystem;
-
-/* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ArmAuto extends Command {
-
-  private ArmSubsystem armSub;
-  private double goalAngle;
-
-  /** Creates a new ArmAuto. */
-  public ArmAuto(ArmSubsystem armSub, double goalAngle) {
-
-    this.armSub = armSub;
-    this.goalAngle = goalAngle;
-
-    // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(armSub);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    armSub.setReferenceAngle(goalAngle);
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
