+++ src/main/java/frc/robot/commands/arm/ArmPID.java
@@ -1,0 +1,32 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.commands.arm;
+
+import edu.wpi.first.wpilibj2.command.InstantCommand;
+import frc.robot.subsystems.arm.ArmSubsystem;
+
+// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
+// information, see:
+// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
+public class ArmPID extends InstantCommand {
+
+  private ArmSubsystem armSub;
+  private double goalAngle;
+
+  public ArmPID(ArmSubsystem armSub, double goalAngle) {
+
+    this.armSub = armSub;
+    this.goalAngle = goalAngle;
+
+    // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(armSub);
+  }
+
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {
+    armSub.setReferenceAngle(goalAngle);
+  }
+}
